LayoutTests/imported/w3c/web-platform-tests/svg/animations/repeat-iteration-event-006-expected.txt


FAIL Animation triggers on 'repeatEvent' iteration assert_equals: Expected repeatEvent event, but got beginEvent event instead expected "repeatEvent" but got "beginEvent"
